import model with texture and sponza

+++ src/scene.cpp
@@ -44,7 +44,9 @@ std::vector<std::shared_ptr<Light>> Scene::GetLights() const {
 }
 
 void Scene::SetActiveSceneObject(unsigned int id) {
-    m_ActiveObjectID = id < SceneObject::GetIDCount() ? id : -1;
+    m_ActiveObjectID = static_cast<int>(id) < SceneObject::GetIDCount()
+                           ? static_cast<int>(id)
+                           : -1;
 }
 
 std::shared_ptr<SceneObject> Scene::GetActiveSceneObject() const {
@@ -52,7 +54,8 @@ std::shared_ptr<SceneObject> Scene::GetActiveSceneObject() const {
     return m_ActiveObjectID != -1
                ? *std::find_if(m_SceneObjects.begin(), m_SceneObjects.end(),
                                [this](std::shared_ptr<SceneObject> obj) {
-                                   return obj->GetID() == m_ActiveObjectID;
+                                   return static_cast<int>(obj->GetID()) ==
+                                          m_ActiveObjectID;
                                })
                : nullptr;
 }
